# backend/cmd/server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for managing financial data, chat interactions, and scenario planning",
        "title": "Financial Chat System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/v2/assets": {
            "get": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Returns paginated non-cash assets for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets V2"
                ],
                "summary": "List assets (v2)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Max items to return (-1 for all)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Creates a new non-cash asset for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets V2"
                ],
                "summary": "Create an asset (v2)",
                "parameters": [
                    {
                        "description": "Asset data",
                        "name": "asset",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.assetCreateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repository.NonCashAsset"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Bulk deletes all non-cash assets for the authenticated user in a single query.",
                "tags": [
                    "Bulk Delete V2"
                ],
                "summary": "Delete all assets (v2)",
                "operationId": "deleteBulkAssets",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v2/assets/{id}": {
            "put": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Updates an asset with versioning support",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets V2"
                ],
                "summary": "Update an asset (v2)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Asset data",
                        "name": "asset",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.assetInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repository.NonCashAsset"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Deletes an asset and its descendant versions",
                "tags": [
                    "Assets V2"
                ],
                "summary": "Delete an asset (v2)",
                "operationId": "deleteAsset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v2/assets/{id}/stop": {
            "post": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Sets the endDate on an asset (soft delete)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets V2"
                ],
                "summary": "Stop an asset (v2)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Stop input with endDate",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.stopInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repository.NonCashAsset"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.assetCreateInput": {
            "type": "object",
            "properties": {
                "annualGrowthRate": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "currentValue": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "growthStrategy": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "terminalValue": {
                    "type": "string"
                }
            }
        },
        "handlers.assetInput": {
            "type": "object",
            "properties": {
                "annualGrowthRate": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "currentValue": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "growthStrategy": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "parentId": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "terminalValue": {
                    "type": "string"
                },
                "updateMode": {
                    "type": "string"
                }
            }
        },
...
        "handlers.stopInput": {
            "type": "object",
            "properties": {
                "endDate": {
                    "type": "string"
                }
            }
        },
...
        "repository.NonCashAsset": {
            "type": "object",
            "properties": {
                "annualGrowthRate": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "currentValue": {
                    "type": "number"
                },
                "endDate": {
                    "description": "NULL means ongoing",
                    "type": "string"
                },
                "growthStrategy": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "impactFrequency": {
                    "description": "How often delta adds (NULL for base/override)",
                    "type": "string"
                },
                "impactKind": {
                    "description": "NULL = base item, 'delta' = additive, 'override' = replaces",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "parentId": {
                    "type": "string"
                },
                "scenarioEventId": {
                    "description": "Scenario impact fields",
                    "type": "string"
                },
                "startDate": {
                    "description": "Precise start date (day-level)",
                    "type": "string"
                },
                "terminalValue": {
                    "description": "Value at end of useful life (NULL = disappear, 0 = worthless)",
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AuthToken": {
            "description": "JWT token for authentication (HMAC-signed with BACKEND_SHARED_SECRET). In dev mode, use GET /api/dev/token?user_id=\u003cuuid\u003e to generate one.",
            "type": "apiKey",
            "name": "X-Auth-Token",
            "in": "header"
        }
    }
}